requirements.txt
torchinfo==0.1.1
matplotlib==3.3.4
numpy==1.19.2
torchvision==0.9.0
fake_useragent==0.1.11
torch==1.8.0
requests==2.26.0
tqdm==4.62.2
beautifulsoup4==4.10.0
Pillow==8.4.0
